requirements.txt
Django
djangorestframework
python-dotenv
sentence-transformers
faiss-cpu
requests
numpy
youtube-transcript-api
